ad9694_500ebz: Allow to configure number of lanes, number of converters and samplerate
To make it easier to select the converter configuration parameters add
top-level parameters to the project script. These are used to calculate
other dependent configuration settings.

This means that if any of the converter settings are changed the dependent
settings are updated automatically.

This makes it easier to try different configurations.

TBD: Figure out how to only have one definition of NUM_LANES rather than three.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>