ad9694_500ebz: Remove transceiver memory map arbitration
The avl_adxcfg core performs arbitration between the register map for the
RX and TX channel of the transceivers. In this project only RX transceivers
are used so arbitration is not necessary. In addition it seems not
connecting the second Avalon-MM port of the avl_adxcfg causes it to
function incorrectly and reads/writes on the single connected port are
lost.

So remove the arbitration core and directly connect the transceiver
register map to the CPU interconnect.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>