ad9694_500ebz: Gate SPI MISO signals based on chip-select
ORing together the two MISO signals can cause problems due to interference.
On the SPI header where the chipselect signal is not asserted the state of
the MISO signal is undefined and may contain bogus values which when
combined with the valid signal might cause it to become corrupted.

Make sure to only use the MISO signal from the header on which the
chip-select is asserted.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>